update: main, command, config, power; broken: sudo creates config in project root - find way to write to files

+++ internal/config/config.go
@@ -18,8 +18,8 @@ import (
 
 type Profile struct {
 	Name  string `json:"name"`
-	Start int64  `json:"start"`
-	End   int64  `json:"end"`
+	Start int    `json:"start"`
+	End   int    `json:"end"`
 }
 
 type Profiles map[string]Profile
